# packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Stealth Console API
  version: 0.1.0
  description: Control-plane API for Stealth Console. Browser sessions are opaque HttpOnly cookies and are never returned in JSON.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/projects/{projectID}/webhooks:
    get:
      operationId: listWebhooks
      description: List webhook delivery configurations without exposing signing secrets. Console owners/admins or project API keys with webhooks.read may read them.
      security: [{ sessionCookie: [] }, { projectApiKey: [] }]
      parameters: [{ $ref: '#/components/parameters/ProjectID' }, { $ref: '#/components/parameters/Limit' }, { $ref: '#/components/parameters/Cursor' }]
      responses:
        '200': { description: Webhook configurations, content: { application/json: { schema: { $ref: '#/components/schemas/WebhooksPage' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      operationId: createWebhook
      description: Create an HTTPS webhook. The whsec_ signing secret is returned exactly once; store it in a server-side secret manager.
      security: [{ sessionCookie: [] }, { projectApiKey: [] }]
      parameters: [{ $ref: '#/components/parameters/ProjectID' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateWebhookRequest' } } } }
      responses:
        '201': { description: Webhook metadata and one-time signing secret, content: { application/json: { schema: { $ref: '#/components/schemas/WebhookSecretResponse' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /v1/projects/{projectID}/webhooks/{webhookID}:
    get:
      operationId: getWebhook
      security: [{ sessionCookie: [] }, { projectApiKey: [] }]
      parameters: [{ $ref: '#/components/parameters/ProjectID' }, { $ref: '#/components/parameters/WebhookID' }]
      responses:
        '200': { description: Webhook metadata, content: { application/json: { schema: { $ref: '#/components/schemas/WebhookResponse' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updateWebhook
      security: [{ sessionCookie: [] }, { projectApiKey: [] }]
      parameters: [{ $ref: '#/components/parameters/ProjectID' }, { $ref: '#/components/parameters/WebhookID' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UpdateWebhookRequest' } } } }
      responses:
        '200': { description: Updated webhook metadata, content: { application/json: { schema: { $ref: '#/components/schemas/WebhookResponse' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      operationId: deleteWebhook
      security: [{ sessionCookie: [] }, { projectApiKey: [] }]
      parameters: [{ $ref: '#/components/parameters/ProjectID' }, { $ref: '#/components/parameters/WebhookID' }]
      responses:
        '204': { description: Webhook and its pending deliveries deleted }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/projects/{projectID}/webhooks/{webhookID}/rotate-secret:
    post:
      operationId: rotateWebhookSecret
      description: Replace the signing secret and return the new whsec_ value exactly once. Existing deliveries use the new secret when they are sent.
      security: [{ sessionCookie: [] }, { projectApiKey: [] }]
      parameters: [{ $ref: '#/components/parameters/ProjectID' }, { $ref: '#/components/parameters/WebhookID' }]
      responses:
        '200': { description: Updated webhook metadata and one-time signing secret, content: { application/json: { schema: { $ref: '#/components/schemas/WebhookSecretResponse' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: stealth_session }
    projectAppCookie: { type: apiKey, in: cookie, name: stealth_app_<project-id-without-hyphens>, description: Project-specific opaque HttpOnly cookie. The actual name is derived from the path projectID; it is never the Console stealth_session cookie. }
    projectApiKey: { type: apiKey, in: header, name: X-Stealth-Key, description: Project-bound server-to-server secret. Supported scopes are users.read/write, databases.read/write, storage.read/write, functions.read/write, sites.read/write, webhooks.read/write, and realtime.read; never send this header through the Console browser bridge. }
  parameters:
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
    Cursor: { name: cursor, in: query, schema: { type: string, format: uuid } }
...
    ProjectID: { name: projectID, in: path, required: true, schema: { type: string, format: uuid } }
...
    WebhookID: { name: webhookID, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
    BadRequest: { description: Invalid JSON or parameter, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    Unauthorized: { description: Authentication required, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
    Forbidden: { description: Insufficient organization role, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    Conflict: { description: Resource conflict, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    NotFound: { description: Resource not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
    InternalError: { description: Internal server error, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
    ServiceUnavailable: { description: Public Auth protection or a required dependency is unavailable, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    PayloadTooLarge: { description: Request body exceeds the endpoint's configured limit (JSON requests are bounded separately from multipart uploads), content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    UnsupportedMediaType: { description: Content-Type must be application/json except for file upload multipart/form-data endpoints, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    ValidationError: { description: Validation failed, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
  schemas:
...
    Pagination: { type: object, required: [limit, next_cursor], properties: { limit: { type: integer }, next_cursor: { type: [string, 'null'], format: uuid } } }
    ErrorEnvelope: { type: object, additionalProperties: false, required: [error], properties: { error: { type: object, additionalProperties: false, required: [code, message], properties: { code: { type: string }, message: { type: string } } } } }
...
    Webhook: { type: object, additionalProperties: false, required: [id, project_id, name, url, events, enabled, failure_count, last_delivery_at, last_failure_at, created_at, updated_at], properties: { id: { type: string, format: uuid }, project_id: { type: string, format: uuid }, name: { type: string, minLength: 2, maxLength: 120 }, url: { type: string, format: uri, pattern: '^https://', maxLength: 2048 }, events: { type: array, minItems: 1, maxItems: 64, uniqueItems: true, items: { type: string, pattern: '^(\*|[a-z0-9][a-z0-9._-]{0,159})$' } }, enabled: { type: boolean }, failure_count: { type: integer, minimum: 0 }, last_delivery_at: { type: [string, 'null'], format: date-time }, last_failure_at: { type: [string, 'null'], format: date-time }, created_at: { type: string, format: date-time }, updated_at: { type: string, format: date-time } } }
    CreateWebhookRequest: { type: object, additionalProperties: false, required: [name, url], properties: { name: { type: string, minLength: 2, maxLength: 120 }, url: { type: string, format: uri, pattern: '^https://', maxLength: 2048, description: HTTPS endpoint; private, loopback, and link-local destinations are blocked during delivery. }, events: { type: array, maxItems: 64, uniqueItems: true, items: { type: string, pattern: '^(\*|[a-z0-9][a-z0-9._-]{0,159})$' }, default: ['*'] }, enabled: { type: boolean, default: true } } }
    UpdateWebhookRequest: { type: object, additionalProperties: false, minProperties: 1, properties: { name: { type: string, minLength: 2, maxLength: 120 }, url: { type: string, format: uri, pattern: '^https://', maxLength: 2048 }, events: { type: array, minItems: 1, maxItems: 64, uniqueItems: true, items: { type: string, pattern: '^(\*|[a-z0-9][a-z0-9._-]{0,159})$' } }, enabled: { type: boolean } } }
    WebhookSecretResponse: { type: object, additionalProperties: false, required: [webhook, secret], properties: { webhook: { $ref: '#/components/schemas/Webhook' }, secret: { type: string, writeOnly: true, pattern: '^whsec_[A-Za-z0-9_-]{43}$', description: High-entropy signing secret shown only by create or rotate. } } }
    WebhookResponse: { type: object, additionalProperties: false, required: [webhook], properties: { webhook: { $ref: '#/components/schemas/Webhook' } } }
...
    WebhooksPage: { type: object, additionalProperties: false, required: [webhooks, pagination, can_manage], properties: { webhooks: { type: array, items: { $ref: '#/components/schemas/Webhook' } }, pagination: { $ref: '#/components/schemas/Pagination' }, can_manage: { type: boolean } } }
...